worker/requirements.txt
protobuf
grpcio
grpcio-tools
requests
sentence-transformers
numpy
